chore: reorg react utils
