--- packages/react-utils/src/react-utils.ts
@@ -1,124 +1,0 @@
-import * as React from "react"
-import * as ReactDOM from "react-dom"
-import { isObject, isFunction, AnyFunction } from "@chakra-ui/utils"
-
-// Event assertions
-export function isInputEvent(value: any): value is React.ChangeEvent {
-  return value && isObject(value) && isObject(value.target)
-}
-
-export interface CreateContextOptions {
-  /**
-   * If `true`, React will throw if context is `null` or `undefined`
-   * In some cases, you might want to support nested context, so you can set it to `false`
-   */
-  strict?: boolean
-  /**
-   * Error message to throw if the context is `undefined`
-   */
-  errorMessage?: string
-  /**
-   * The display name of the context
-   */
-  name?: string
-}
-
-type CreateContextReturn<T> = [React.Provider<T>, () => T, React.Context<T>]
-
-/**
- * Creates a named context, provider, and hook.
- *
- * @param options create context options
- */
-export function createContext<ContextType>(options: CreateContextOptions = {}) {
-  const {
-    strict = true,
-    errorMessage = "useContext: `context` is undefined. Seems you forgot to wrap component within the Provider",
-    name,
-  } = options
-
-  const Context = React.createContext<ContextType | undefined>(undefined)
-
-  Context.displayName = name
-
-  function useContext() {
-    const context = React.useContext(Context)
-
-    if (!context && strict) {
-      throw new Error(errorMessage)
-    }
-
-    return context
-  }
-
-  return [
-    Context.Provider,
-    useContext,
-    Context,
-  ] as CreateContextReturn<ContextType>
-}
-
-/**
- * Gets only the valid children of a component,
- * and ignores any nullish or falsy child.
- *
- * @param children the children
- */
-export function getValidChildren(children: React.ReactNode) {
-  return React.Children.toArray(children).filter((child) =>
-    React.isValidElement(child),
-  ) as React.ReactElement[]
-}
-
-type ReactRef<T> = React.Ref<T> | React.RefObject<T> | React.MutableRefObject<T>
-
-/**
- * Assigns a value to a ref function or object
- *
- * @param ref the ref to assign to
- * @param value the value
- */
-export function assignRef<T = any>(ref: ReactRef<T> | undefined, value: T) {
-  if (ref == null) return
-
-  if (isFunction(ref)) {
-    ref(value)
-    return
-  }
-
-  try {
-    // @ts-ignore
-    ref.current = value
-  } catch (error) {
-    throw new Error(`Cannot assign value '${value}' to ref '${ref}'`)
-  }
-}
-
-/**
- * Combine multiple React refs into a single ref function.
- * This is used mostly when you need to allow consumers forward refs to
- * internal components
- *
- * @param refs refs to assign to value to
- */
-export function mergeRefs<T>(...refs: (ReactRef<T> | undefined)[]) {
-  return (value: T) => {
-    refs.forEach((ref) => assignRef(ref, value))
-  }
-}
-
-/** This is a workaround for React Concurrent Mode issue https://github.com/facebook/react/issues/18591. Remove once it's fixed. */
-export function withFlushSync<T extends AnyFunction>(fn: any) {
-  return (event: Parameters<T>[0]) => {
-    const flushSync = (ReactDOM as any).flushSync as
-      | ((fn: AnyFunction) => void)
-      | undefined
-    if (isFunction(flushSync)) {
-      flushSync(() => {
-        fn(event)
-      })
-    } else {
-      fn(event)
-    }
-  }
-}
